CDPCP-2673. Poll status of refresh when syncing OIDs
When setting cloud identity settings in ranger, the SDX service
now also returns a status object that the caller can use to
poll the status of the sync (which is just the status of the underlying
CM refresh command). The FreeIPA service polls this object to verify
that the sync is successful.

In order to simplify the design, the cloud identity sync code
has been modified to only work for single datalake per environment
and fail if there are multiple datalakes present (which we currently
don't support anyway).