CB-7729 Retrieve errors from Data Lake Database backup/restore failures
Pulls the status reason for failed database backup/restore operations from the
salt stderr ouput, instead of from the command comment. The is achieved by
adding a new YAML file used by SaltErrorResolver, stderrcommands.yaml. When a
command listed in that file is run, the stderr output will be used as the
resolved error message. The getErrorResultSummary() method in RunnerInfo now
returns the summary in map form for simpler parsing in SaltErrorResolver.

Tested with unit tests and by testing in my local cloudbreak dev environment.